Revert calculator v2 and redirects; keep aesthetic updates
- Restore pre-v2 DumpsterCalculator.tsx
- Remove DumpsterCalculatorV2.tsx and DumpsterCalculatorLegacy.tsx
- Remove feature flag system (lib/featureFlags.ts)
- Revert vercel.json redirects to previous working config
